refactor a bit more and get rid of os.mkdir
